+++ packages/core/CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+[Unreleased]
+
+### Changed
+
+### Added
+
+- Added support to allow the scout driver to be defined per model.
+
+### Fixed
+
 ## 2.0-beta12 - 2022-04-08
 
 ### Changed